t/signal/senales/ closes #44
