+++ examples/senales/señales.sh
@@ -2,7 +2,11 @@
 # signal. By typing `ctrl-C` (which the
 # terminal shows as `^C`) we can send a `SIGINT` signal,
 # causing the program to print `interrupt` and then exit.
-$ go run signals.go
+# Cuando corremos este programa se bloqueara 
+# esperando una señal. Al escribir `ctrl-C` (que se 
+# mostrará en la terminal como `^C`) podemos enviar
+# la señal `SIGINT`.
+$ go run señales.go
 awaiting signal
 ^C
 interrupt
