+++ public/index.html
@@ -161,7 +161,7 @@
       
         <li><a href="reemplazando-procesos-con-exec">Reemplazando procesos con exec</a></li>
       
-        <li><a href="signals">Signals</a></li>
+        <li><a href="senales">Señales</a></li>
       
         <li><a href="exit">Exit</a></li>
