+++ public/reemplazando-procesos-con-exec
@@ -23,8 +23,8 @@
         
         <tr>
           <td class="docs">
-            <p>En el previo ejemplo vimos como
-<a href="creaando-procesos">crear un nuevo proceso</a>. Hacemos
+            <p>En el ejemplo previo vimos como
+<a href="craando-procesos">crear un nuevo proceso</a>. Hacemos
 esto cuando necesitamos que el proceso sea accesible
 al proceso en Go que se está ejecutando. A veces
 necesitamos reemplazar completamente el proceso
@@ -34,7 +34,7 @@ función <a href="http://en.wikipedia.org/wiki/Exec_(operating_system)"><code>ex
 
           </td>
           <td class="code leading">
-          <a href="http://play.golang.org/p/StAW_-LFp1"><img title="Run code" src="play.png" class="run" /></a>
+          <a href="http://play.golang.org/p/VRK-0FaHCB"><img title="Run code" src="play.png" class="run" /></a>
             <div class="highlight"><pre><span class="kn">package</span> <span class="nx">main</span>
 </pre></div>
 
@@ -104,7 +104,7 @@ debe de ser el nombre del programa.</p>
         
         <tr>
           <td class="docs">
-            <p><code>Exec</code> también necesita un conjunto de <a href="variable-de-entorno">variables de entorno</a>.
+            <p><code>Exec</code> también necesita un conjunto de <a href="variables-de-entorno">variables de entorno</a>.
 Aquí solamente le pasaremos nuestro ambiente
 actual.</p>
 
@@ -180,7 +180,7 @@ cubre la mayoría de los usos de <code>fork</code>.</p>
       
       
       <p class="next">
-        Siguiente ejemplo: <a href="signals">Signals</a>.
+        Siguiente ejemplo: <a href="senales">Señales</a>.
       </p>
       
       <p class="footer">
